Refactor: moved encoding.js and sweetalert to libs section as well
