+++ docs/index.html
@@ -14,10 +14,12 @@
     <!-- libs -->
     <script type="text/javascript" src="./js/libs/CanvasRenderingContext2D.ext.js"></script>
     <script type="text/javascript" src="./js/libs/easytimer.min.js"></script>
+    <script type="text/javascript" src="./js/libs/encoding.js"></script>
     <script type="text/javascript" src="./js/libs/jquery-3.5.1.min.js"></script>
     <script type="text/javascript" src="./js/libs/timer.js"></script>
     <script type="text/javascript" src="./js/libs/vanillaSelectBox.js"></script>
     <script type="text/javascript" src="./js/libs/zlib.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@10"></script>
     <!-- rest -->
     <script type="text/javascript" src="./js/main.js"></script>
     <script type="text/javascript" src="./js/class_p.js"></script>
@@ -29,8 +31,6 @@
     <script type="text/javascript" src="./js/class_panel.js"></script>
     <script type="text/javascript" src="./js/style.js"></script>
     <script type="text/javascript" src="./js/general.js"></script>
-    <script type="text/javascript" src="./js/encoding.js"></script>
-    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@10"></script>
     <!-- Global site tag (gtag.js) - Google Analytics -->
     <script async src="https://www.googletagmanager.com/gtag/js?id=UA-180924087-1"></script>
     <script>
